requirements.txt
queue
copy
floor
matplotlib.pyplot 
numpy 
pandas
random
csv
pprint

